+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Emitted `Event::TrackChanged` events now contains full metadata.
 - Compilation warnings caused by newer Rust versions (up to 1.28) have been
   fixed.
+- `Player::set_volume` is fixed (always set to 0 previously)
 
 ### Added